Entered code for computerSelection function
The code entered will return a choice based off of a random number, as
stated in the comments for the code. Also, created a variable that and
logged it to the console for testing purposes. The function seems to
work as expected, returning a random element from the choices array.