Altered playerSelection function structure
This function now takes in a string value and then stores a version that
was cast into lower case form into a variable. A while loops will be
used to check if the input exists as an option inside the array of
choices. The loop will repeat until an acceptable input is given by the
user. This assures that a valid value is being compared to the computer
choice.