Added styles to scoreBox div and tags within
The scoreBox amd it's content, which include, the span, winScore,
and lossScore, were modified. the scoreBox div was given a border,
padding, and margins for the top. The span was given some padding
and a larger font size to give some space between the score and the
win/loss text. The win/lose score are given a specific color.