Created function and variables to update score
A function, updateScore, will be used to update the current score. The
The winDisplay and lossDisplay variables grants us access to the html
section containing the winning and losing scores. Two other variables,
winScore and lossScore, are used to store the running scores and are
incremented by 1 within the playRound function. Those two variables are
passed as parameters for the updateScore function, which is called
within the playRound function as well.