Created function that alters css based on result
The function, addResultDisplay, allows the css to be modified by adding
or removing the appropriate class based on the outcome of the round.
This changes the result text to blue, if the player won, red, of the
player lost, or black, if it was a draw. The lose/win css styles are
added or removed from the resultBox div based on the outcome.