Fixed logic error within addResultStyle function
The addResultStyle function was not working as intended. Depending on
the order in which a win/loss occurred, the previous class and it's font
color altering property would remain in place. For example, a win,
resulting in blue text, followed by a loss would still display as blue,
instead of its desired red font color. The previous class would remain
along with the new class. The issue was resolved by removing both class
styles at the very beginning of the function, essentially resetting the
resultBox div.